+++ index.js
@@ -1,0 +1,15 @@
+module.exports = {
+
+  // Rule options:
+  //  'off'   / 0
+  //  'warn'  / 1
+  //  'error' / 2
+
+  extends: [
+    './rules/logic.js',
+    './rules/syntax.js',
+    './rules/whitespace.js',
+    './rules/es6.js'
+  ]
+
+};