+++ CHANGELOG.md
@@ -1,20 +1,42 @@
-<a name="1.1.0"></a>
-# [1.1.0](https://github.com/uglow/test-repo/compare/v1.0.0...v1.1.0) (2016-04-13)
+<a name="1.2.0"></a>
+# [1.2.0](https://bitbucket.cd.auspost.com.au/projects/npm/repos/eslint-config-auspost/compare/diff?targetBranch=refs%2Ftags%2Fv1.1.2&sourceBranch=refs%2Ftags%2Fv1.2.0) (2017-01-19)
 
 
 ### Features
 
-* **2:** 2 ([9520377](https://github.com/uglow/test-repo/commit/9520377))
+* **rules:** add no-unused-vars rule ([7c2e23b](https://bitbucket.cd.auspost.com.au/projects/npm/repos/eslint-config-auspost/commits/7c2e23b))
 
+
+<a name="1.1.2"></a>
+# [1.1.2](https://bitbucket.cd.auspost.com.au/projects/npm/repos/eslint-config-auspost/compare/diff?targetBranch=refs%2Ftags%2Fv1.1.1&sourceBranch=refs%2Ftags%2Fv1.1.2) (2017-01-10)
+
+### Fixes
+
+* Changed space-before-function-paren rule to be more specific
+
+
+<a name="1.1.1"></a>
+# [1.1.1](https://bitbucket.cd.auspost.com.au/projects/npm/repos/eslint-config-auspost/compare/diff?targetBranch=refs%2Ftags%2Fv1.1.0&sourceBranch=refs%2Ftags%2Fv1.1.1) (2016-10-25)
+
+### Fixes
+
+* Removed unrecognised rule - symbol-description
+* Relegated prefer-arrow-callback and duplicate-imports to warnings
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://bitbucket.cd.auspost.com.au/projects/npm/repos/eslint-config-auspost/compare/diff?targetBranch=refs%2Ftags%2Fv1.0.0&sourceBranch=refs%2Ftags%2Fv1.1.0) (2016-10-21)
+
+### Features
+
+* Separated rules out into logical groups to aid comprehension / editing
 
 
 <a name="1.0.0"></a>
-# 1.0.0 (2016-04-13)
-
+# [1.0.0](2016-08-01)
 
 ### Features
 
-* **1:** first change ([ccff051](https://github.com/uglow/test-repo/commit/ccff051))
-
+* Initial setup